Move CUDA tests to test/cuda/ and add detailed diagnostics
Move cuda_test.cpp, benchmark_cuda.cpp, and run_cuda_test.sh into
test/cuda/ for cleaner separation from the CPU test suite.

Enhance cuda_test.cpp with per-flavor, per-energy, per-zenith CPU vs
GPU comparison output and summary statistics (mean, max absolute and
relative error per flavor) to help identify divergent regimes.

Edited-with: Claude <noreply@anthropic.com>
